Initial commit of metaclass based api extension
By adding a metaclass to SeekerView, we can control the creation of each
subclass, altering and adding class variables as we see fit. This is
perfect for the extension of an API, because we can keep the old API of class
variable for configuration, and create a new system that can will be translated
into the old one when the class is created.